# Conflicts:
#	elza/dao-test-bench/pom.xml
#	elza/elza-core/pom.xml
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/ApController.java
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/ArrangementController.java
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/FundController.java
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/SearchController.java
#	elza/elza-core/src/main/java/cz/tacr/elza/controller/factory/ApFactory.java
#	elza/elza-core/src/main/java/cz/tacr/elza/dataexchange/output/DEExportService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/dataexchange/output/filters/AccessRestrictFilter.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/ApIndex.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/ApKeyValue.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/ApPart.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/ApState.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/ArrDataUriRef.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/ArrFund.java
#	elza/elza-core/src/main/java/cz/tacr/elza/domain/bridge/ApCachedAccessPointClassBridge.java
#	elza/elza-core/src/main/java/cz/tacr/elza/print/OutputModel.java
#	elza/elza-core/src/main/java/cz/tacr/elza/repository/ApAccessPointRepositoryImpl.java
#	elza/elza-core/src/main/java/cz/tacr/elza/repository/ApCachedAccessPointRepositoryImpl.java
#	elza/elza-core/src/main/java/cz/tacr/elza/security/ApplicationSecurity.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/AccessPointDataService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/AccessPointService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/ArrangementService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/AsyncRequestService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/DescriptionItemService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/ExternalSystemService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/FundLevelService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/RuleService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/arrangement/DeleteFundHistoryAction.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/cache/AccessPointCacheService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/cam/CamService.java
#	elza/elza-core/src/main/java/cz/tacr/elza/service/output/AsyncOutputGeneratorWorker.java
#	elza/elza-core/src/test/java/cz/tacr/elza/AbstractTest.java
#	elza/elza-core/src/test/java/cz/tacr/elza/controller/AbstractControllerTest.java
#	elza/elza-web/src/main/java/cz/tacr/elza/web/controller/ElzaWebController.java
#	elza/elza-ws-api/pom.xml
#	elza/pom.xml
#	elza/rules-simple-dev/pom.xml

+++ elza/elza-core/src/main/java/cz/tacr/elza/service/ExternalSystemService.java
@@ -1,20 +1,35 @@
 package cz.tacr.elza.service;
 
+import java.time.OffsetDateTime;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
 import java.util.Objects;
 import java.util.Optional;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import jakarta.annotation.Nullable;
 import jakarta.transaction.Transactional;
+import jakarta.transaction.Transactional.TxType;
 import jakarta.validation.constraints.NotNull;
 
 import org.apache.commons.lang3.Validate;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 import org.springframework.util.StringUtils;
 
+import cz.tacr.cam.schema.cam.EntityRecordRevInfoXml;
+import cz.tacr.elza.api.ApExternalSystemType;
 import cz.tacr.elza.common.ObjectListIterator;
 import cz.tacr.elza.core.data.StaticDataService;
 import cz.tacr.elza.core.security.AuthMethod;
@@ -22,6 +37,7 @@ import cz.tacr.elza.domain.ApAccessPoint;
 import cz.tacr.elza.domain.ApBinding;
 import cz.tacr.elza.domain.ApBindingItem;
 import cz.tacr.elza.domain.ApBindingState;
+import cz.tacr.elza.domain.ApBindingSync;
 import cz.tacr.elza.domain.ApChange;
 import cz.tacr.elza.domain.ApExternalSystem;
 import cz.tacr.elza.domain.ApItem;
@@ -29,21 +45,29 @@ import cz.tacr.elza.domain.ApPart;
 import cz.tacr.elza.domain.ApType;
 import cz.tacr.elza.domain.ArrDigitalRepository;
 import cz.tacr.elza.domain.ArrDigitizationFrontdesk;
+import cz.tacr.elza.domain.ExtSyncsQueueItem;
+import cz.tacr.elza.domain.ExtSyncsQueueItem.ExtAsyncQueueState;
 import cz.tacr.elza.domain.SyncState;
 import cz.tacr.elza.domain.SysExternalSystem;
 import cz.tacr.elza.domain.UsrPermission;
+import cz.tacr.elza.domain.UsrPermission.Permission;
+import cz.tacr.elza.domain.UsrUser;
+import cz.tacr.elza.domain.enumeration.StringLength;
 import cz.tacr.elza.exception.BusinessException;
 import cz.tacr.elza.exception.ObjectNotFoundException;
 import cz.tacr.elza.exception.SystemException;
 import cz.tacr.elza.exception.codes.BaseCode;
+import cz.tacr.elza.repository.ApAccessPointRepository;
 import cz.tacr.elza.repository.ApBindingItemRepository;
 import cz.tacr.elza.repository.ApBindingRepository;
 import cz.tacr.elza.repository.ApBindingStateRepository;
+import cz.tacr.elza.repository.ApBindingSyncRepository;
 import cz.tacr.elza.repository.ApExternalSystemRepository;
 import cz.tacr.elza.repository.DigitalRepositoryRepository;
 import cz.tacr.elza.repository.DigitizationFrontdeskRepository;
 import cz.tacr.elza.repository.ExtSyncsQueueItemRepository;
 import cz.tacr.elza.repository.ExternalSystemRepository;
+import cz.tacr.elza.service.cam.BindingSyncInfo;
 import cz.tacr.elza.service.eventnotification.events.EventId;
 import cz.tacr.elza.service.eventnotification.events.EventType;
 
@@ -54,6 +78,11 @@ import cz.tacr.elza.service.eventnotification.events.EventType;
  */
 @Service
 public class ExternalSystemService {
+
+    static private final Logger log = LoggerFactory.getLogger(ExternalSystemService.class);
+
+    @Autowired
+    private ApAccessPointRepository apAccessPointRepository;
 
     @Autowired
     ExtSyncsQueueItemRepository extSyncsQueueItemRepository;
@@ -77,6 +106,9 @@ public class ExternalSystemService {
     private ApBindingRepository bindingRepository;
 
     @Autowired
+    private ApBindingSyncRepository bindingSyncRepository;
+
+    @Autowired
     private ApBindingStateRepository bindingStateRepository;
 
     @Autowired
@@ -84,6 +116,9 @@ public class ExternalSystemService {
 
     @Autowired
     private StaticDataService staticDataService;
+
+    @Autowired
+    private UserService userService;
 
     /**
      * Vyhledá všechny externí systémy.
@@ -211,9 +246,27 @@ public class ExternalSystemService {
      *
      * @param extSyncItemId
      */
-    @AuthMethod(permission = UsrPermission.Permission.ADMIN)
     public void deleteQueueItem(final Integer extSyncItemId) {
-        extSyncsQueueItemRepository.deleteById(extSyncItemId);
+        Optional<ExtSyncsQueueItem> queueItem = extSyncsQueueItemRepository.findById(extSyncItemId);
+
+        if (queueItem.isPresent()) {
+
+            ExtSyncsQueueItem item = queueItem.get();
+            UsrUser loggedUser = userService.getLoggedUser();
+
+            // smazat záznam může pouze Admin nebo autor
+            if (!userService.hasPermission(Permission.ADMIN)) {
+                if (loggedUser == null || !Objects.equals(loggedUser.getUserId(), item.getUserId())) {
+                    throw new SystemException("Uživatel nemá oprávnění na vymazávání přístupového bodu ve frontě", BaseCode.INSUFFICIENT_PERMISSIONS)
+                    .set("accessPointId", item.getAccessPointId())
+                    .set("userId", item.getUserId());
+                }
+            }
+
+            if (queueItem.get().getState() != ExtAsyncQueueState.EXPORT_START) {
+                extSyncsQueueItemRepository.deleteById(extSyncItemId);
+            }
+        }
     }
 
     /**
@@ -416,6 +469,7 @@ public class ExternalSystemService {
         apBindingState.setApExternalSystem(binding.getApExternalSystem());
         apBindingState.setExtState(state);
         apBindingState.setExtRevision(revisionUuid);
+        Validate.isTrue(userName == null || userName.length() <= StringLength.LENGTH_250, "UserName length exceeds the limit");
         apBindingState.setExtUser(userName);
         apBindingState.setExtReplacedBy(replacedById == null ? null : String.valueOf(replacedById));
         apBindingState.setSyncChange(apChange);
@@ -423,7 +477,8 @@ public class ExternalSystemService {
         apBindingState.setSyncOk(syncState);
         apBindingState.setPreferredPart(preferredPart);
         apBindingState.setApType(apType);
-        return bindingStateRepository.save(apBindingState);
+
+        return bindingStateRepository.saveAndFlush(apBindingState);
     }
 
     /**
@@ -463,21 +518,16 @@ public class ExternalSystemService {
         oldbindingState.setDeleteChange(apChange);
         bindingStateRepository.saveAndFlush(oldbindingState);
 
-        ApBindingState apBindingState = new ApBindingState();
-        apBindingState.setBinding(oldbindingState.getBinding());
-        apBindingState.setAccessPoint(oldbindingState.getAccessPoint());
-        apBindingState.setApExternalSystem(oldbindingState.getApExternalSystem());
-        apBindingState.setExtState(state);
-        apBindingState.setExtRevision(revisionUuid);
-        apBindingState.setExtUser(user);
-        apBindingState.setExtReplacedBy(extReplacedBy);
-        apBindingState.setSyncChange(apChange);
-        apBindingState.setCreateChange(apChange);
-        apBindingState.setSyncOk(syncState);
-        apBindingState.setPreferredPart(preferredPart);
-        apBindingState.setApType(apType);
-
-        return bindingStateRepository.saveAndFlush(apBindingState);
+        return createBindingState(oldbindingState.getBinding(),
+                                  oldbindingState.getAccessPoint(),
+                                  apChange,
+                                  state,
+                                  revisionUuid,
+                                  user,
+                                  extReplacedBy == null? null : Long.valueOf(extReplacedBy),
+                                  syncState,
+                                  preferredPart,
+                                  apType);
     }
 
     public ApBindingItem createApBindingItem(final ApBinding binding,
@@ -512,6 +562,137 @@ public class ExternalSystemService {
         return bindingStateRepository.findActiveByBinding(binding);
     }
 
+    @Transactional
+    public BindingSyncInfo getBindingSync(final String extSystemCode, final String transactionUuid) {
+        ApExternalSystem externalSystem = findApExternalSystemByCode(extSystemCode);
+
+        ApBindingSync bindingSync = bindingSyncRepository.findByApExternalSystem(externalSystem);
+        if (bindingSync == null) {
+            bindingSync = new ApBindingSync();
+            bindingSync.setApExternalSystem(externalSystem);
+            bindingSync.setLastTransaction(transactionUuid);
+            bindingSync = bindingSyncRepository.save(bindingSync);
+        }
+        return new BindingSyncInfo(bindingSync.getBindingSyncId(),
+                                   externalSystem.getExternalSystemId(),
+                                   bindingSync.getLastTransaction(), bindingSync.getToTransaction(),
+                                   bindingSync.getPage(), bindingSync.getCount());
+    }
+
+    /**
+     * Prepare entities for synchronization
+     *
+     * @param bindingSyncId
+     * @param entityRecordRevInfoXmls entity info list
+     * @param lastTransaction
+     * @param toTransaction
+     * @param page
+     * @param count
+     */
+    @Transactional
+    public void prepareApsForSync(Integer bindingSyncId, List<EntityRecordRevInfoXml> entityRecordRevInfoXmls,
+                                  String lastTransaction, String toTransaction,
+                                  Integer page, Integer count) {
+        log.debug("Preparing APs for synchronization from external system, count: {}", entityRecordRevInfoXmls.size());
+
+        // Prepare keys
+        ApBindingSync bindingSync = bindingSyncRepository.getOneCheckExist(bindingSyncId);
+        ApExternalSystem externalSystem = bindingSync.getApExternalSystem();
+        List<String> keyList = new ArrayList<>(entityRecordRevInfoXmls.size());
+        Map<String, EntityRecordRevInfoXml> recordCodesMap = new HashMap<>();
+        Function<EntityRecordRevInfoXml, String> idGetter;
+        if (externalSystem.getType().equals(ApExternalSystemType.CAM_UUID)) {
+            idGetter = (x) -> x.getEuid().getValue();
+        } else {
+            idGetter = (x) -> Long.toString(x.getEid().getValue());
+        }
+        for (EntityRecordRevInfoXml entityRecordRevInfoXml : entityRecordRevInfoXmls) {
+            String id = idGetter.apply(entityRecordRevInfoXml);
+            keyList.add(id);
+            EntityRecordRevInfoXml prevInfo = recordCodesMap.put(id, entityRecordRevInfoXml);
+            Validate.isTrue(prevInfo == null, "Record with same key already process, %s", id);
+        }
+
+        List<ApBinding> bindings = findBindings(keyList, externalSystem);
+        final Map<String, ApBinding> bindingMap = bindings.stream().collect(Collectors.toMap(p -> p.getValue(), p -> p));
+
+        Map<Integer, ApBindingState> bindingStateMap;
+        if (bindings.size() > 0) {
+            List<ApBindingState> bindingStateList = findBindingStates(bindings);
+            bindingStateMap = bindingStateList.stream().collect(Collectors.toMap(p -> p.getBindingId(), p -> p));
+        } else {
+            bindingStateMap = Collections.emptyMap();
+        }
+
+        int recNo = 0;
+
+        UsrUser user = userService.getLoggedUser();
+        for (String recordCode : keyList) {
+            recNo++;
+            if (log.isDebugEnabled()) {
+                if (recNo%100 == 0) {
+                    log.debug("Prepared records for sync: [{}-{}]", ((recNo+99)/100-1)*100+1, recNo);
+                }
+            }
+
+            ApBinding binding = bindingMap.get(recordCode);
+            ApAccessPoint ap = null;
+            if (binding == null) {
+                // prepare binding for CAM Complete
+                if (externalSystem.getType() == ApExternalSystemType.CAM_COMPLETE) {
+                    // we are creating all bindings at once
+                    // - will be flush to the DB at the end of this method
+                    binding = createApBinding(recordCode, externalSystem, false);
+                }
+            } else {
+                ApBindingState bindingState = bindingStateMap.get(binding.getBindingId());
+                EntityRecordRevInfoXml xmlRecordInfo = recordCodesMap.get(recordCode);
+                if (bindingState != null) {
+                    ap = bindingState.getAccessPoint();
+                    // kontrola uuid revizi, pokud se rovná extRevizion(), pak aktualizace není potřeba
+                    String uuidRev = xmlRecordInfo.getRev().getValue();
+                    if (bindingState.getExtRevision().equals(uuidRev)) {
+                        continue;
+                    }
+                }
+                // entita mohla být smazána, hledáme ji jinak
+                if (ap == null) {
+                    String uuid = xmlRecordInfo.getEuid().getValue();
+                    ap = apAccessPointRepository.findAccessPointByUuid(uuid);
+                }
+            }
+            // update or add new items from CAM_COMPLETE
+            if (ap != null || externalSystem.getType() == ApExternalSystemType.CAM_COMPLETE) {
+                createExtSyncsQueueItem(ap, externalSystem, binding, null,
+                                        ap != null? ExtAsyncQueueState.UPDATE : ExtAsyncQueueState.IMPORT_NEW,
+                                        OffsetDateTime.now(),
+                                        user);
+            }
+        }
+        if (log.isDebugEnabled()) {
+            if (recNo%100 != 0) {
+                log.debug("Prepared records for sync: [{}-{}]", ((recNo+99)/100-1)*100+1, recNo);
+            }
+        }
+        log.debug("APs prepared for synchronization from external system");
+        log.info("To queue ext_syncs_queue_item added {} records for sync.", recNo);
+
+        // aktualizace dat
+        bindingSync.setLastTransaction(lastTransaction);
+        bindingSync.setToTransaction(toTransaction);
+        bindingSync.setPage(page);
+        bindingSync.setCount(count);
+        bindingSyncRepository.saveAndFlush(bindingSync);
+    }
+
+    @Transactional
+    public void resetTransaction(final Integer bindingSyncId, final String transactionUuid) {
+        ApBindingSync bindingSync = bindingSyncRepository.getOneCheckExist(bindingSyncId);
+        bindingSync.setLastTransaction(transactionUuid);
+        bindingSync.setPage(null);
+        bindingSyncRepository.save(bindingSync);
+    }
+
     public ApBindingItem findByBindingAndUuid(ApBinding binding, String uuid) {
         return bindingItemRepository.findByBindingAndUuid(binding, uuid);
     }
@@ -544,4 +725,54 @@ public class ExternalSystemService {
         return ObjectListIterator.findIterable(bindings,
                                                bs -> bindingStateRepository.findByBindings(bs));
     }
-}
+
+    /**
+     * Vytvoření záznamu ve frontě zpracování
+     *
+     * @param accessPoint
+     * @param apExternalSystem
+     * @param stateMessage
+     * @param state
+     * @param date
+     * @param user
+     * @return
+     */
+    public ExtSyncsQueueItem createExtSyncsQueueItem(final ApAccessPoint accessPoint,
+                                                     final ApExternalSystem apExternalSystem,
+                                                     final ApBinding binding,
+                                                     final String stateMessage,
+                                                     final ExtSyncsQueueItem.ExtAsyncQueueState state,
+                                                     final OffsetDateTime date,
+                                                     final UsrUser user) {
+         ExtSyncsQueueItem extSyncsQueueItem = new ExtSyncsQueueItem();
+         extSyncsQueueItem.setAccessPoint(accessPoint);
+         extSyncsQueueItem.setExternalSystem(apExternalSystem);
+         extSyncsQueueItem.setBinding(binding);
+         extSyncsQueueItem.setStateMessage(stateMessage);
+         extSyncsQueueItem.setState(state);
+         extSyncsQueueItem.setDate(date);
+         extSyncsQueueItem.setUser(user);
+
+         return extSyncsQueueItemRepository.save(extSyncsQueueItem);
+     }
+
+     /**
+      * Return list of first items to process in given states
+      *
+      * @param pageSize
+      * @param states
+      * @return
+      */
+     @Transactional(value = TxType.MANDATORY)
+     public Iterable<ExtSyncsQueueItem> getNextItems(int pageSize, ExtAsyncQueueState... states) {
+         Pageable pageable = PageRequest.of(0, pageSize);
+
+         return extSyncsQueueItemRepository.findByStates(Arrays.asList(states), pageable);
+     }
+
+     public List<ExtSyncsQueueItem> getQueueItems(Integer accessPointId, Integer externalSystemId,
+                                                  ExtAsyncQueueState... states) {
+         return extSyncsQueueItemRepository.findByApExtSystAndState(accessPointId, externalSystemId, Arrays.asList(
+                                                                                                                   states));
+     }
+ }
